# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobRecord": {
        "additionalProperties": true,
        "description": "A row of ``GET /api/jobs`` (one dataset view record).\n\nExtra columns are preserved so future SQL fields do not break the response\nmodel; the declared fields mirror the current query projection.",
        "properties": {
          "ai_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Reason"
          },
          "ai_score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Score"
          },
          "company": {
            "default": "",
            "title": "Company",
            "type": "string"
          },
          "date_posted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Posted"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "evaluated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluated At"
          },
          "first_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "gemini_gaps": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Gaps"
          },
          "gemini_interview_chance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Interview Chance"
          },
          "gemini_matched_skills": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Matched Skills"
          },
          "gemini_score": {
            "default": 0,
            "title": "Gemini Score",
            "type": "integer"
          },
          "gemini_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Status"
          },
          "gemini_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gemini Summary"
          },
          "job_url": {
            "default": "",
            "title": "Job Url",
            "type": "string"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "matched_skills": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Skills"
          },
          "score": {
            "default": 0,
            "title": "Score",
            "type": "integer"
          },
          "scraped_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scraped At"
          },
          "target_archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Archetype"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "JobRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Karriere Pipeline Dashboard",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/job-descriptions": {
      "get": {
        "description": "Resolve job descriptions by URL, falling back to (company, position).",
        "operationId": "job_descriptions_api_job_descriptions_get",
        "parameters": [
          {
            "in": "query",
            "name": "urls",
            "required": false,
            "schema": {
              "default": "",
              "title": "Urls",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "company",
            "required": false,
            "schema": {
              "default": "",
              "title": "Company",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "position",
            "required": false,
            "schema": {
              "default": "",
              "title": "Position",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Descriptions",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs": {
      "get": {
        "description": "Return job records for a named SQLite view.",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "file",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "File"
            }
          },
          {
            "in": "query",
            "name": "dataset",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobRecord"
                  },
                  "title": "Response List Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}